Commit built library directly.
This change is meant only for internal use at Cohesic while the
status of the upstream repository is in limbo.  A number of PRs
are pending and it appears that the repo may be abandoned.
Should the upstream be abandoned, this project will move to the
Cohesic group and be published under @cohesic on npm.